requirements.txt
fastapi==0.85.1
uvicorn==0.18.3 
gunicorn==20.1.0
langchain==0.1.20
langchain-community==0.0.38
pydantic>=1.6.2
python-dotenv==1.0.1
pypdf==3.16.4
openai==1.28.1
chromadb==0.3.29
tiktoken==0.6.0
numpy==1.26.4
faiss-cpu==1.8.0